refactor: update tests without dictionary
- Use Factory's get instead of return value of read/readall
- Put the test case which demonstrates the extraction to a network storage into test_stream_nwstorage.py
- Move the test case which demonstrates the stream extraction and packing into test_stream.py
