+++ tests/test_basic.py
@@ -272,8 +272,7 @@ def test_py7zr_read_and_reset(tmp_path):
     archive = py7zr.SevenZipFile(open(os.path.join(testdata_path, "read_reset.7z"), "rb"))
     iterations = archive.getnames()
     for target in iterations:
-        _dict = archive.read(factory=py7zr.io.NullIOFactory(), targets=[target])
-        assert len(_dict) == 1
+        archive.read(factory=py7zr.io.NullIOFactory(), targets=[target])
         archive.reset()
     archive.close()
 
@@ -283,8 +282,7 @@ def test_py7zr_read_with_trailing_slash_and_reset(tmp_path):
     archive = py7zr.SevenZipFile(open(os.path.join(testdata_path, "read_reset.7z"), "rb"))
     iterations = archive.getnames()
     for target in iterations:
-        _dict = archive.read(factory=py7zr.io.NullIOFactory(), targets=[f"{target}/"])
-        assert len(_dict) == 1
+        archive.read(factory=py7zr.io.NullIOFactory(), targets=[f"{target}/"])
         archive.reset()
     archive.close()
 
@@ -350,14 +348,13 @@ def test_read_collection_argument():
     )
     factory = py7zr.io.BytesIOFactory(64)
     with py7zr.SevenZipFile(BytesIO(data), password="boom") as arc:
-        result = arc.read(factory, ["bar.txt"])  # list -> ok
-        assert "bar.txt" in result
-        bina = result.get("bar.txt")
-        assert isinstance(bina, MemIO)
+        arc.read(factory, ["bar.txt"])  # list -> ok
+        assert "bar.txt" in factory.products
+        bina = factory.get("bar.txt")
         assert bina.read() == b"refinery"
     with py7zr.SevenZipFile(BytesIO(data), password="boom") as arc:
-        result = arc.read(factory, {"bar.txt"})  # set -> ok
-        assert result.get("bar.txt").read() == b"refinery"
+        arc.read(factory, {"bar.txt"})  # set -> ok
+        assert factory.get("bar.txt").read() == b"refinery"
     with pytest.raises(TypeError):
         with py7zr.SevenZipFile(BytesIO(data), password="boom") as arc:
             arc.read(factory, ("bar.txt",))  # tuple -> bad
