+++ tests/test_extract.py
@@ -50,10 +50,11 @@ def check_archive(archive, tmp_path, return_dict: bool):
         assert tmp_path.joinpath("test/test2.txt").open("rb").read() == bytes("This file is located in a folder.", "ascii")
         assert tmp_path.joinpath("test1.txt").open("rb").read() == bytes("This file is located in the root.", "ascii")
     else:
-        _dict = archive.readall(py7zr.io.BytesIOFactory(64))
-        actual = _dict["test/test2.txt"].read()
+        factory = py7zr.io.BytesIOFactory(64)
+        archive.readall(factory)
+        actual = factory.get("test/test2.txt").read()
         assert actual == bytes("This file is located in a folder.", "ascii")
-        actual = _dict["test1.txt"].read()
+        actual = factory.get("test1.txt").read()
         assert actual == bytes("This file is located in the root.", "ascii")
     archive.close()
 
@@ -91,8 +92,9 @@ def test_github_14(tmp_path):
 @pytest.mark.files
 def test_github_14_mem(tmp_path):
     archive = py7zr.SevenZipFile(testdata_path.joinpath("github_14.7z").open(mode="rb"))
-    _dict = archive.readall(py7zr.io.BytesIOFactory(limit=32))
-    actual = _dict["github_14"].read()
+    factory = py7zr.io.BytesIOFactory(limit=32)
+    archive.readall(factory)
+    actual = factory.get("github_14").read()
     assert actual == bytes("Hello GitHub issue #14.\n", "ascii")
 
 
@@ -105,8 +107,9 @@ def _test_umlaut_archive(filename: str, target: pathlib.Path, return_dict: bool)
         actual = target.joinpath("t\xe4st.txt").open().read()
         assert actual == "This file contains a german umlaut in the filename."
     else:
-        _dict = archive.readall(py7zr.io.BytesIOFactory(64))
-        actual = _dict["t\xe4st.txt"].read()
+        factory = py7zr.io.BytesIOFactory(64)
+        archive.readall(factory)
+        actual = factory.products["t\xe4st.txt"].read()
         assert actual == b"This file contains a german umlaut in the filename."
     archive.close()
 
@@ -248,8 +251,9 @@ def test_lzma2bcj_mem():
         "mingw64/share/doc/szip/RELEASE.txt",
         "mingw64/bin/libszip-0.dll",
     ]
-    _dict = archive.readall(py7zr.io.HashIOFactory())
-    digest = _dict["mingw64/bin/libszip-0.dll"].read()
+    factory = py7zr.io.HashIOFactory()
+    archive.readall(factory)
+    digest = factory.products["mingw64/bin/libszip-0.dll"].read()
     assert digest == binascii.unhexlify("13926e3f080c9ca557165864ce5722acc4f832bb52a92d8d86c7f6e583708c4d")
     archive.close()
 
@@ -285,7 +289,7 @@ def test_extract_lzma_1(tmp_path):
 def test_extract_lzma2_1(tmp_path):
     with testdata_path.joinpath("lzma2_1.7z").open(mode="rb") as target:
         with py7zr.SevenZipFile(target) as ar:
-            _dict = ar.readall(py7zr.io.NullIOFactory())
+            ar.readall(py7zr.io.NullIOFactory())
 
 
 @pytest.mark.files
@@ -358,10 +362,11 @@ def test_github_14_multi_mem():
     """multiple unnamed objects."""
     archive = py7zr.SevenZipFile(str(testdata_path.joinpath("github_14_multi.7z")), "r")
     assert archive.getnames() == ["github_14_multi", "github_14_multi"]
-    _dict = archive.readall(py7zr.io.BytesIOFactory(32))
-    actual_1 = _dict["github_14_multi"].read()
+    factory = py7zr.io.BytesIOFactory(32)
+    archive.readall(factory)
+    actual_1 = factory.get("github_14_multi").read()
     assert actual_1 == bytes("Hello GitHub issue #14 1/2.\n", "ascii")
-    actual_2 = _dict["github_14_multi_0"].read()
+    actual_2 = factory.get("github_14_multi_0").read()
     assert actual_2 == bytes("Hello GitHub issue #14 2/2.\n", "ascii")
     archive.close()
 
@@ -379,8 +384,9 @@ def test_multiblock(tmp_path):
 @pytest.mark.files
 def test_multiblock_mem():
     archive = py7zr.SevenZipFile(testdata_path.joinpath("mblock_1.7z").open(mode="rb"))
-    _dict = archive.readall(py7zr.io.HashIOFactory())
-    digest = _dict["bin/7zdec.exe"].read()
+    factory = py7zr.io.HashIOFactory()
+    archive.readall(factory)
+    digest = factory.products["bin/7zdec.exe"].read()
     assert digest == binascii.unhexlify("e14d8201c5c0d1049e717a63898a3b1c7ce4054a24871daebaa717da64dcaff5")
     archive.close()
 
@@ -602,14 +608,11 @@ def test_extract_root_path_arcname(tmp_path):
         iterations = archive.getnames()
         assert len(iterations) == 1
 
-        _dict = archive.read(factory=py7zr.io.BytesIOFactory(32), targets=iterations)
-        if _dict is None:
-            # fix typing errors
-            raise RuntimeError("Failed to read archive")
-
-        assert len(_dict) == 1
-        assert [*_dict.keys()] == [filename]
-        assert _dict[filename].read() == content
+        factory = py7zr.io.BytesIOFactory(32)
+        archive.read(factory=factory, targets=iterations)
+        assert len(factory.products) == 1
+        assert filename in factory.products.keys()
+        assert factory.products[filename].read() == content
 
     with py7zr.SevenZipFile(filename_7z, "r") as archive:
         archive.extractall(path=tmp_path)
